use unique id for all variables instead of only temporaries
The previous approach, which renamed variables in inlining was invalid.
nested scope variables were not handled (using a hierarchical visitor
would fix that) and loop unrolling too would result on duplicated
variables with same name. I encountred same variables names on output
even after fixing these two problems, so I guess there is more cases
where a variable is duplicated and it's name needs to be changed but is
not.

This alternative solution, which append a unique id to each variable
seems more robust and future proof (should iteract well with future
optimization, without needing to check if it duplicates variables).